backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
bidict==0.23.1
build==1.4.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==0.6.3
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
fastapi==0.135.1
filelock==3.25.0
firebase_admin==7.2.0
flatbuffers==25.12.19
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-aiplatform==1.139.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.65.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
limits==5.8.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
posthog==5.4.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.5
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.1
wsproto==1.3.2
zipp==3.23.0
